Conflicts:
	target/classes/com/yami/action/user/UserAction.class
	target/classes/com/yami/aop/InertAspect.class
	target/classes/com/yami/mapper/UserMapper.class
	target/classes/com/yami/service/UserService.class
	target/classes/com/yami/service/impl/UserServiceImpl.class
	target/classes/mybatis-config/sqlmap-config.xml
	target/test-classes/com/yami/test/user/UserServiceImplTest.class

+++ src/main/java/com/yami/service/UserService.java
@@ -1,6 +1,6 @@
 package com.yami.service;
 
-import com.yami.dto.User;
+import com.yami.domain.User;
 
 public interface UserService {
 	
@@ -8,5 +8,5 @@ public interface UserService {
 	
 	User getUserByUserId(String userId);
 	
-	int insertUser(User user);
+	long insertUser(User user);
 }
